about to deploy to live

+++ server/controllers/supplierController.js
@@ -1,3 +1,4 @@
+import ProductModel from '../models/ProductModel.js';
 import Supplier from '../models/SupplierModel.js'
 
 const addSupplier = async (req, res) => {
@@ -51,24 +52,32 @@ const updateSupplier = async (req, res) => {
     return res.status(200).json({ success: true, message: 'Supplier updated successfully!' });
   } catch (error) {
     console.error('Error updating supplier:', error);
-    return res.status(500).json({success: false, message:'Server error'})
-  
+    return res.status(500).json({ success: false, message: 'Server error' })
+
   }
 }
 
-const deleteDelete = async (req, res) =>{
+const deleteDelete = async (req, res) => {
   try {
-    const {id} = req.params;
-  const existingSupplier = await Supplier.findById(id);
-  if (!existingSupplier) {
-    return res.status(404).json({ success: false, message: ' Supplier not found'})
-  }
+    const { id } = req.params;
 
-  await Supplier.findByIdAndDelete(id);
-  return res.status(200).json({success: true, message: 'Supplier deleted successfully'})
+    const productCount = await ProductModel.countDocuments({ supplierId: id })
+
+    if (productCount > 0) {
+      return res.status(400).json({ success: false, message: "Can not delete supplier associated with products" })
+    }
+
+
+    const existingSupplier = await Supplier.findById(id);
+    if (!existingSupplier) {
+      return res.status(404).json({ success: false, message: ' Supplier not found' })
+    }
+
+    await Supplier.findByIdAndDelete(id);
+    return res.status(200).json({ success: true, message: 'Supplier deleted successfully' })
   } catch (error) {
     console.error('Error deleting supplier', error);
-    return res.status(500).json({success: false, message: 'Server error'})
+    return res.status(500).json({ success: false, message: 'Server error' })
   }
 }
